clean ファイル整理
