--- 08.ls_object/lib/viewer.rb
@@ -1,6 +1,0 @@
-#!/usr/bin/env ruby
-# frozen_string_literal: true
-
-module Viewer
-
-end
